requirements.txt
beautifulsoup4==4.12.3
lxml==5.3.0
sentence-transformers==3.1.1
torch>=2.2.0
qdrant-client==1.9.2
elasticsearch==8.15.1
neo4j==5.22.0
python-dotenv==1.0.1
PyPDF2==3.0.1
tqdm==4.66.4
numpy>=1.26.0
google-generativeai==0.8.3
openai>=1.55.0
tenacity==9.0.0